Move .gitignore handling out of init.ts into its own module
init.ts previously changed for two unrelated reasons: the Q&A/stub
pipeline and gitignore glob matching. Extracting the latter into
gitignore.ts keeps each module changing for one reason and gives the
gitignore logic its own seam. Also covers the leading-slash-anchored
form with a test and tightens the ADR's wording on what the wildcard
check actually supports.
