+++ docs/adr/0026-init-gitignores-readyrun-dir.md
@@ -1,3 +1,3 @@
 # Init auto-appends .readyrun/ to the Consumer's .gitignore
 
-`run` creates `.readyrun/worktrees/<branch>` inside the Consumer's repo root (ADR 0008), but nothing ReadyRun ships gitignored it — a Consumer had to know to add `.readyrun/` themselves or risk committing Worktree contents by accident. `init` now ensures the Consumer's `.gitignore` covers `.readyrun/` as its last step, after writing the config stub: if `.gitignore` doesn't exist, `init` creates one containing just that line; if it exists and lacks the line, `init` appends it, preserving whatever is already there. Before appending, `init` checks whether an existing line already covers the directory — either the exact `.readyrun/` (with or without the leading `/` anchor or trailing slash) or a broader glob against the same basename, such as `.ready*` or `.*` — so a Consumer who already ignores it a different way doesn't get a redundant duplicate line. This check is deliberately conservative: a pattern anchored to a nested path (containing a `/` other than a trailing one) is treated as not covering the root-level directory, since fully resolving arbitrary gitignore path semantics is out of scope for a one-line convenience check.
+`run` creates `.readyrun/worktrees/<branch>` inside the Consumer's repo root (ADR 0008), but nothing ReadyRun ships gitignored it — a Consumer had to know to add `.readyrun/` themselves or risk committing Worktree contents by accident. `init` now ensures the Consumer's `.gitignore` covers `.readyrun/` as its last step, after writing the config stub: if `.gitignore` doesn't exist, `init` creates one containing just that line; if it exists and lacks the line, `init` appends it, preserving whatever is already there. Before appending, `init` checks whether an existing line already covers the directory — either the exact `.readyrun/` (with or without the leading `/` anchor or trailing slash) or a `*`/`?` wildcard against the same basename, such as `.ready*` or `.*` — so a Consumer who already ignores it a different way doesn't get a redundant duplicate line. This check is deliberately conservative and does not implement full gitignore glob semantics: a pattern anchored to a nested path (containing a `/` other than a trailing one), or using gitignore's `[...]` character-class syntax, is treated as not covering the root-level directory rather than risked as a false positive.
